% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_captions_from_json}
\alias{extract_captions_from_json}
\title{Extract available languages in captions}
\usage{
extract_captions_from_json(json_l, x)
}
\arguments{
\item{json_l}{A list object, resulting from a parsed .info.json file}

\item{x}{Type of data to extract. In practice, either "subtitles" or
"automatic_captions"}
}
\value{
A list, with names of available languages
}
\description{
Used internally
}
\examples{
\dontrun{
extract_captions_from_json(
  yyjsonr::read_json_file("path_to_json_file"),
  "automatic_captions"
)
}
}
