% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yt_check_local_subtitles.R
\name{yt_check_local_subtitles}
\alias{yt_check_local_subtitles}
\title{Check for the availability of local subtitles and return details in a data
frame}
\usage{
yt_check_local_subtitles(sub_format = "vtt", yt_base_folder = NULL)
}
\arguments{
\item{sub_format}{Defaults to "vtt". File extension of the subtitle.}

\item{yt_base_folder}{Base folder, defaults to NULL. Can be set with
\code{\link[=yt_set_base_folder]{yt_set_base_folder()}}}
}
\value{
A data frame (a tibble) with details on locally available subtitle
files.
}
\description{
Check for the availability of local subtitles and return details in a data
frame
}
\examples{
\dontrun{
yt_check_local_subtitles()
}
}
