% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yt_concatenate.R
\name{yt_concatenate}
\alias{yt_concatenate}
\title{Concatenated trimmed video files in a single video clip}
\usage{
yt_concatenate(
  trimmed_df,
  sort_by_date = FALSE,
  info_df = NULL,
  destination_filename = "concatenated",
  destination_folder = "0_concatenated_video",
  destination_path = NULL,
  overwrite = FALSE,
  yt_base_folder = NULL
)
}
\arguments{
\item{trimmed_df}{A data frame, typically generated as output of \code{yt_trim()} or \code{yt_trim_with_text()}.}

\item{sort_by_date}{Defaults to FALSE. If TRUE, retrieves (and, if necessary, downloads) info files for all video clips, and reads them with \code{yt_read_info_json()} in order to get information about the published time.}

\item{info_df}{Defaults to NULL. If given, a data frame typically generated with \code{yt_read_info_json()}.}

\item{destination_folder}{Defaults to \verb{0_trimmed_video}: trimmed video files
will be stored inside a folder called \verb{0_trimmed_video} inside your base
folder as defined by \code{yt_base_folder} or options. If you want an absolute
path, use the argument \code{destination_path} instead.}

\item{destination_path}{Defaults to NULL. Location where trimmed video files
will be stored. If given, takes precedence over \code{destination_folder}.}

\item{overwrite}{Defaults to FALSE. If TRUE, overwrites concatenated file if it already exists.}

\item{yt_base_folder}{Base folder, defaults to NULL. Can be set with
\code{\link[=yt_set_base_folder]{yt_set_base_folder()}}}
}
\value{
The path of the generated video clip.
}
\description{
Concatenated trimmed video files in a single video clip
}
