% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yt_extract_id.R
\name{yt_extract_id}
\alias{yt_extract_id}
\title{Extract YouTube identifier from an URL.}
\usage{
yt_extract_id(yt_id)
}
\arguments{
\item{yt_id}{Url or unique identifier of a YouTube video.}
}
\value{
A character vector of YouTube identifiers.
}
\description{
Extract YouTube identifier from an URL.
}
\examples{
yt_extract_id("https://youtu.be/WXPBOfRtXQE?feature=shared")

# if already an identifier, just returns it:
yt_extract_id("WXPBOfRtXQE")
}
