% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yt_get_available_subtitles.R
\name{yt_get_available_subtitles}
\alias{yt_get_available_subtitles}
\title{Checks with clips have subtitles or automatic captions based on info json
file}
\usage{
yt_get_available_subtitles(
  info_json_df = NULL,
  sub_lang = "en",
  automatic_captions = TRUE,
  subtitles = TRUE,
  yt_id = NULL,
  playlist = NULL
)
}
\arguments{
\item{info_json_df}{Defaults to NULL. Generally created with
\code{yt_read_info_json()}. If not given, either \code{yt_id} or \code{playlist} must be
given. If given, both \code{yt_id} and \code{playlist} are ignored.}

\item{sub_lang}{Defaults to "en", subtitles language.}

\item{automatic_captions}{Defaults to TRUE. If TRUE, checks if subtitles are
available as "automatic captions".}

\item{subtitles}{Defaults to TRUE. If TRUE, checks if subtitles are available
as (manually added or approved) "subtitles".}

\item{yt_id}{YouTube video identifier. Ignored if \code{info_json_df} given.}

\item{playlist}{YouTube list. Ignored if \code{info_json_df} given.}
}
\value{
A data frame with three columns, \code{yt_id}, \code{sub_lang}, and \code{sub_type}
(\code{sub_type} can either be \code{automatic_captions} or \code{subtitles}).
}
\description{
Checks with clips have subtitles or automatic captions based on info json
file
}
\examples{
\dontrun{
yt_get_available_subtitles(
  playlist = "https://www.youtube.com/playlist?list=PLbyvawxScNbtMcDKmT2dRAfjmSFwOt1Vj"
)
}
}
