% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yt_set_base_folder.R
\name{yt_get_base_folder}
\alias{yt_get_base_folder}
\title{Retrieves base folder where all retrieved files will be stored for the
current session}
\usage{
yt_get_base_folder(path = NULL)
}
\arguments{
\item{path}{Defaults to NULL. If not set, checks if a path has been
previously set, and if not, defaults to current folder.}
}
\value{
The given path (or, if left to NULL, the path previously set) is
returned invisibly.
}
\description{
Retrieves base folder where all retrieved files will be stored for the
current session
}
\examples{
yt_set_base_folder(path = fs::path(
  fs::path_home_r(),
  "R",
  "ytdlpr"
))
yt_get_base_folder()
}
