% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yt_check_local_subtitles.R
\name{yt_get_local_subtitles}
\alias{yt_get_local_subtitles}
\title{Check for the availability of local subtitles and return details in a data
frame}
\usage{
yt_get_local_subtitles(
  playlist = NULL,
  language = NULL,
  sub_format = "vtt",
  yt_base_folder = NULL
)
}
\arguments{
\item{playlist}{Playlist, either as full url from Youtube or as id.}

\item{language}{Defaults to NULL. If not given, all local subtitles are
returned. If given, only subtitles in the given language are returned.}

\item{sub_format}{Defaults to "vtt". File extension of the subtitle.}

\item{yt_base_folder}{Base folder, defaults to NULL. Can be set with
\code{\link[=yt_set_base_folder]{yt_set_base_folder()}}}
}
\value{
A data frame (a tibble) with details on locally available subtitle
files.
}
\description{
Check for the availability of local subtitles and return details in a data
frame
}
\examples{
\dontrun{
yt_get_local_subtitles()
}
}
