% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yt_get_playlist_folder.R
\name{yt_get_playlist_folder}
\alias{yt_get_playlist_folder}
\title{Get local path to folder where files for the given playlist will be stored}
\usage{
yt_get_playlist_folder(playlist, yt_base_folder = NULL)
}
\arguments{
\item{playlist}{Playlist, either as full url from Youtube or as id.}

\item{yt_base_folder}{Base folder, defaults to NULL. Can be set with
\code{\link[=yt_set_base_folder]{yt_set_base_folder()}}}
}
\value{
Path to playlist folder.
}
\description{
Get local path to folder where files for the given playlist will be stored
}
\examples{
\dontrun{
yt_get_playlist_folder(
  playlist = "https://www.youtube.com/playlist?list=PLbyvawxScNbtMcDKmT2dRAfjmSFwOt1Vj"
)
}

}
