% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yt_get_playlist_id.R
\name{yt_get_playlist_id}
\alias{yt_get_playlist_id}
\title{Get all Youtube identifiers for a given playlist}
\usage{
yt_get_playlist_id(playlist, update = FALSE, yt_base_folder = NULL)
}
\arguments{
\item{playlist}{The full url of a Youtube playlist.}

\item{update}{Defaults to FALSE. If FALSE, data is returned immediately if
previously stored. If TRUE, it checks again the playlist on Youtube to see
if new content has been added.}

\item{yt_base_folder}{Base folder, defaults to NULL. Can be set with
\code{\link[=yt_set_base_folder]{yt_set_base_folder()}}}
}
\value{
A data frame (a tibble) with a single column named \code{yt_id}.
}
\description{
Get all Youtube identifiers for a given playlist
}
\examples{
\dontrun{
yt_get_playlist_id(
  playlist = "https://www.youtube.com/playlist?list=PLbyvawxScNbtMcDKmT2dRAfjmSFwOt1Vj"
)
}
}
