% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yt_read_vtt.R
\name{yt_read_vtt}
\alias{yt_read_vtt}
\title{Read vtt subtitles}
\usage{
yt_read_vtt(path)
}
\arguments{
\item{path}{Path to one or more subtitle file in the vtt format. If a data frame is used as input, a column named "path" in that data frame will be used as source.}
}
\value{
A data frame with
}
\description{
Read vtt subtitles
}
\examples{
\dontrun{
yt_get(
  playlist = "https://www.youtube.com/playlist?list=PLbyvawxScNbuSi7sJaJbHNyyx3iYJeW3P",
  auto_subs = TRUE
) |>
  yt_read_vtt()
}
}
