% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yt_set_base_folder.R
\name{yt_set_base_folder}
\alias{yt_set_base_folder}
\title{Set for the current session base folder where all retrieved files will be
stored}
\usage{
yt_set_base_folder(path = NULL)
}
\arguments{
\item{path}{Defaults to NULL. If not set, checks if a path has been
previously set, and if not, defaults to current folder.}
}
\value{
The given path (or, if left to NULL, the path previously set) is
returned invisibly.
}
\description{
Set for the current session base folder where all retrieved files will be
stored
}
\examples{
yt_set_base_folder(path = fs::path(
  fs::path_home_r(),
  "R",
  "ytdlpr"
))
yt_get_base_folder()
}
