% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yt_trim.R
\name{yt_trim}
\alias{yt_trim}
\title{Trim video files so that it shows only the part with relevant subtitles}
\usage{
yt_trim(
  subtitles_df,
  lag = -3,
  duration = 6,
  video_file_extension = "webm|mp4|mkv",
  simulate = FALSE,
  destination_folder = "0_trimmed_video",
  destination_path = NULL,
  yt_base_folder = NULL
)
}
\arguments{
\item{subtitles_df}{A data frame with subtitles, typically generated with a
combination of \code{\link[=yt_read_vtt]{yt_read_vtt()}} and \code{\link[=yt_filter]{yt_filter()}}. See examples.}

\item{lag}{Defaults to \code{-3}. Refers to the number of seconds before or after
the start time as recorded in the subtitles. Minus three or four seems to
generally be a good fit.}

\item{duration}{Duration in seconds of the trimmed video. Defaults to 6.}

\item{simulate}{Defaults to FALSE. Similiarly to the same argument in
\code{yt-dlp}, if set to TRUE nothing is actually done}

\item{destination_folder}{Defaults to \verb{0_trimmed_video}: trimmed video files
will be stored inside a folder called \verb{0_trimmed_video} inside your base
folder as defined by \code{yt_base_folder} or options. If you want an absolute
path, use the argument \code{destination_path} instead.}

\item{destination_path}{Defaults to NULL. Location where trimmed video files
will be stored. If given, takes precedence over \code{destination_folder}.}

\item{yt_base_folder}{Base folder, defaults to NULL. Can be set with
\code{\link[=yt_set_base_folder]{yt_set_base_folder()}}}
}
\value{
A data fram with details about the export and ffmpeg commmand. Mostly
used for side effects (creates trimmed video files).
}
\description{
Trim video files so that it shows only the part with relevant subtitles
}
\examples{
\dontrun{
filtered_subs_df <- yt_get(
  yt_id = "-0pPBAiJaYk",
  subtitles = TRUE,
  video = TRUE
) |>
  yt_read_vtt() |>
  yt_filter(pattern = "community")

yt_trim(filtered_subs_df)
}
}
