% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yt_get.R
\name{yt_get}
\alias{yt_get}
\title{Retrieve video or write_subs from a playlist or by id}
\usage{
yt_get(
  yt_id = NULL,
  playlist = NULL,
  check_previous = TRUE,
  sub_lang = "en",
  sub_format = "vtt",
  subs = FALSE,
  auto_subs = FALSE,
  video = FALSE,
  description = FALSE,
  info_json = FALSE,
  comments = FALSE,
  thumbnail = FALSE,
  min_sleep_interval = 1,
  max_sleep_interval = 8,
  sleep_subtitles = 2,
  custom_options = "",
  custom_binary = NULL,
  yt_base_folder = NULL
)
}
\arguments{
\item{yt_id}{YouTube identifier of a video or full url to a video. Normally a
vector, but if a data frame is given, \code{yt_id} column is taken as input
(useful for pipe-based workflows).}

\item{playlist}{Playlist, either as full url from Youtube or as id.}

\item{check_previous}{Defaults to TRUE. If FALSE, input is always downloaded.
If TRUE, and \code{subs} is TRUE, it checks that the requested language is
locally available. If subtitles is set to FALSE, and only one of the write
options is enabled, it checks for the local availabilty of relevant files.
For some options, the presence of \emph{any} local file associated with a given
id prevents further downloads associated with it.}

\item{sub_lang}{Defaults to "en". If more than one, can be given as comma
separated two letter codes, or a as vector.}

\item{sub_format}{Defaults to "vtt". Other formats not yet supported.}

\item{subs}{Defaults to FALSE. "Write subtitle file".}

\item{auto_subs}{Defaults to FALSE "Write automatically generated
subtitle file".}

\item{video}{Defaults to FALSE. Download the video files.}

\item{description}{Defaults to FALSE. "Write video description to a
.description file"}

\item{info_json}{Defaults to FALSE. "Write video metadata to a .info.json
file"}

\item{comments}{Defaults to FALSE. "Retrieve video comments to be placed in
the infojson."}

\item{thumbnail}{Defaults to FALSE. "Write thumbnail image to disk"}

\item{min_sleep_interval}{"Number of seconds to sleep before each download.
This is the minimum time to sleep when used along with
--max-sleep-interval"}

\item{max_sleep_interval}{"Maximum number of seconds to sleep. Can only be
used along with --min-sleep-interval"}

\item{sleep_subtitles}{"Number of seconds to sleep before each subtitle
download"}

\item{custom_options}{Defaults to an empty string. If given, it should
correspond to parameters exactly as they would be used on command line. For
a full list, see the \href{https://github.com/yt-dlp/yt-dlp}{original documentation}.}

\item{custom_binary}{Defaults to NULL. If given, path to executable \code{yt-dlp}.}

\item{yt_base_folder}{Base folder, defaults to NULL. Can be set with
\code{\link[=yt_set_base_folder]{yt_set_base_folder()}}}
}
\value{
A data frame, with details about locally available subtitles.
}
\description{
If \code{write_subs} or \code{auto_subs} is set to TRUE, it checks that the
subtitles in the language set with \code{sub_lang} are available. In all other
cases, by defaults, it skips ' downloads if \emph{any} previous file associated
with a given video identifier has ' been downloaded. Set \code{check_previous} to
FALSE to always download files.
}
\details{
Argument definition are quoted from the original yt-dlp project.
}
\examples{
\dontrun{
yt_get(
  playlist = "https://www.youtube.com/playlist?list=PLbyvawxScNbtMcDKmT2dRAfjmSFwOt1Vj"
)

yt_get(
  yt_id = "https://youtu.be/WXPBOfRtXQE",
  subtitles = TRUE,
  video = TRUE,
  description = TRUE,
  info_json = TRUE
)
}
}
