% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yt_get.R
\name{yt_get}
\alias{yt_get}
\title{Retrieve video or subtitles from a playlist or by id}
\usage{
yt_get(
  yt_id = NULL,
  playlist = NULL,
  subtitles = FALSE,
  sub_lang = "en",
  sub_format = "vtt",
  video = FALSE,
  description = FALSE,
  comments = FALSE,
  info_json = FALSE,
  thumbnail = FALSE,
  write_auto_sub = TRUE,
  min_sleep_interval = 1,
  max_sleep_interval = 8,
  sleep_subtitles = 2,
  custom_options = "",
  yt_base_folder = NULL
)
}
\arguments{
\item{yt_id}{YouTube identifier of a video or full url to a video.}

\item{playlist}{Playlist, either as full url from Youtube or as id.}

\item{sub_lang}{Defaults to "en". If more than one, can be given as comma
separated two letter codes, or a as vector.}

\item{sub_format}{Defaults to "vtt". Other formats not yet supported.}

\item{video}{Defaults to FALSE. Download the video files.}

\item{description}{Defaults to FALSE. "Write video description to a .description file"}

\item{comments}{Defaults to FALSE. "Retrieve video comments to be placed in the infojson."}

\item{info_json}{Defaults to FALSE. "Write video metadata to a .info.json file"}

\item{thumbnail}{Defaults to FALSE. "Write thumbnail image to disk"}

\item{write_auto_sub}{Defaults to TRUE. "Write automatically generated subtitle file"}

\item{min_sleep_interval}{"Number of seconds to sleep before each download.
This is the minimum time to sleep when used along with
--max-sleep-interval"}

\item{max_sleep_interval}{"Maximum number of seconds to sleep. Can only be
used along with --min-sleep-interval"}

\item{sleep_subtitles}{"Number of seconds to sleep before each subtitle
download"}

\item{custom_options}{Defaults to an empty string. If given, it should
correspond to parameters exactly as they would be used on command line. For
a full list, see the \href{https://github.com/yt-dlp/yt-dlp}{original documentation}.}

\item{yt_base_folder}{Base folder, defaults to NULL. Can be set with
\code{\link[=yt_set_base_folder]{yt_set_base_folder()}}}

\item{subtitle}{Defaults to FALSE. "Write subtitle file"}
}
\value{
A data frame, with details about locally available subtitles.
}
\description{
If subtitle is set to TRUE, it checks that the subtitles in the language set with \code{sub_lang} are available. In all other cases, by defaults, it skips downloads if any previous file associated with a given video identifier has been downloaded.
}
\details{
Argument definition are quoted from the original yt-dlp project.
}
\examples{
\dontrun{
yt_get(
  playlist = "https://www.youtube.com/playlist?list=PLbyvawxScNbtMcDKmT2dRAfjmSFwOt1Vj"
)
}
}
