% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yt_get_local_id.R
\name{yt_get_local_id}
\alias{yt_get_local_id}
\title{Lists locally available files that can be attributed to a video id}
\usage{
yt_get_local_id(playlist = NULL, yt_base_folder = NULL)
}
\arguments{
\item{playlist}{Playlist, either as full url from Youtube or as id.}

\item{yt_base_folder}{Base folder, defaults to NULL. Can be set with
\code{\link[=yt_set_base_folder]{yt_set_base_folder()}}}
}
\value{
A data frame, with two columns, \code{yt_id} and \code{path}.
}
\description{
Lists locally available files that can be attributed to a video id
}
\examples{
\dontrun{
yt_get_local_id()
}
}
